@@ -2,9 +2,9 @@ package com.isep.ffa.config;
 
 import com.baomidou.mybatisplus.annotation.DbType;
 import com.baomidou.mybatisplus.extension.plugins.MybatisPlusInterceptor;
-import com.baomidou.mybatisplus.extension.plugins.inner.PaginationInnerInterceptor;
-import com.baomidou.mybatisplus.extension.plugins.inner.OptimisticLockerInnerInterceptor;
-import com.baomidou.mybatisplus.extension.plugins.inner.BlockAttackInnerInterceptor;
+// import com.baomidou.mybatisplus.extension.plugins.inner.PaginationInnerInterceptor;
+// import com.baomidou.mybatisplus.extension.plugins.inner.OptimisticLockerInnerInterceptor;
+// import com.baomidou.mybatisplus.extension.plugins.inner.BlockAttackInnerInterceptor;
 import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -23,14 +23,11 @@ public class MybatisPlusConfig {
   public MybatisPlusInterceptor mybatisPlusInterceptor() {
     MybatisPlusInterceptor interceptor = new MybatisPlusInterceptor();
 
-    // 分页插件
-    interceptor.addInnerInterceptor(new PaginationInnerInterceptor(DbType.POSTGRE_SQL));
-
-    // 乐观锁插件
-    interceptor.addInnerInterceptor(new OptimisticLockerInnerInterceptor());
-
-    // 防止全表更新与删除插件
-    interceptor.addInnerInterceptor(new BlockAttackInnerInterceptor());
+    // 插件配置 - 暂时注释，需要单独引入插件依赖
+    // interceptor.addInnerInterceptor(new
+    // PaginationInnerInterceptor(DbType.POSTGRE_SQL));
+    // interceptor.addInnerInterceptor(new OptimisticLockerInnerInterceptor());
+    // interceptor.addInnerInterceptor(new BlockAttackInnerInterceptor());
 
     return interceptor;
   }
